CDMS Database Diagram
This diagram is a PowerPoint illustration that shows the relationships between the database tables.  While the diagram does show how the database is wired together, it does not include recent changes.  These recent changes are primarily new datasets/tables (_header/_detail), or minor field updates.  However, the main functionality of the database has not changed.